content/canvas/test/test_2d.path.isPointInPath.outside.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.path.isPointInPath.outside</title>
<!-- Testing: isPointInPath() works on paths outside the canvas -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.rect(0, -100, 20, 20);
ctx.rect(20, -10, 20, 20);
ok(ctx.isPointInPath(10, -110) === false, "ctx.isPointInPath(10, -110) === false");
ok(ctx.isPointInPath(10, -90) === true, "ctx.isPointInPath(10, -90) === true");
ok(ctx.isPointInPath(10, -70) === false, "ctx.isPointInPath(10, -70) === false");
ok(ctx.isPointInPath(30, -20) === false, "ctx.isPointInPath(30, -20) === false");
ok(ctx.isPointInPath(30, 0) === true, "ctx.isPointInPath(30, 0) === true");
ok(ctx.isPointInPath(30, 20) === false, "ctx.isPointInPath(30, 20) === false");

SimpleTest.finish();

});
</script>
